Entropy of subsistence strategies for K = 2, 7, 15.
For each cluster from the different K's considered, calculation of the entropy of the mean values of EA001 to EA005 (% of Gathering, Hunting, Fishing, Husbandry and Agriculture). Obtention of the distribution of entropies for each cluster as well. Summary ridgeline plots with the entropy distributions of the clusters obtained for each K.

Inputs:
-"hc_k_2_interpretation.xlsx"
-"hc_k_7_interpretation.xlsx"
-"hc_k_15_interpretation.xlsx"

Scripts calculation of the entropy of the mean values of EA001:EA005:
-"Entropy2.R"
-"Entropy7.R"
-"Entropy15.R"

Scripts calculation of the entropy distributions and obtention of the Ridgeline plots:
-"RidgePlotEntropy2_Vir.R"
-"RidgePlotEntropy7_Vir.R"
-"RidgePlotEntropy15_Vir.R"

Outputs:
1. Tables of entropies:
   -"entropy_2.xlsx"
   -"entropy_7.xlsx"
   -"entropy_15.xlsx"

2. Tables of entropies edited for the paper:
   -"entropy_2_edited.xlsx"
   -"entropy_7_edited.xlsx"
   -"entropy_15_edited.xlsx"

3. Ridgeline plots:
   -RidgePlotEntropy_k2.pdf"
   -RidgePlotEntropy_k7.pdf"
   -RidgePlotEntropy_k15.pdf"